Fixed text selecting to fully show bounds of highlighted text, making it easier for the user to see what they have selected.
Text copied will now blink red. Can be customised to a different colour.

Changed keyboard events so that while control is held, the user will see highlighting for how much space will be consumed if they are to paste.

Added custom colours for highlighted text background and colour. Also custom highlight colour for when control is held.

Allowed the use of Cut & Paste to remember all branches which point to where the user cut from and modify the branches once the selection is pasted so that the branches all point to the new offset location.

Now when modifying the "ADDIU SP, SP" instructions at the beginning or end of a function, all load and store instructions which use SP will be offset so that they still point to the same data after modifying the stack pointer.

The 'RA' part of all shift right arithmetic instructions (eg. 'SRA' or 'DSRA32') will no longer be highlighted when selecting an RA register.